test(input): add tests
